add stub to headers to make libgcc compile

+++ hagen/src/getopt.c
@@ -2,4 +2,4 @@
 int optind = 1;
 int opterr = 1;
 int optopt = 0;
-int optarg = NULL;
+int optarg = (int)NULL;
